requirements.txt
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
